pyproject.toml
[tool.poetry]
name = "alphazero-tetris"
version = "1.0.0"
description = ""
authors = ["Maximilian Weichart <none@none.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jax = "^0.4.34"
tyro = "^0.8.12"
chex = "^0.1.87"
dm-tree = "^0.1.8"
flax = "^0.10.0"
optax = "^0.2.3"
tqdm = "^4.66.5"
tetris-gymnasium = "^0.3.0"
pygraphviz = "^1.14"
orbax = "^0.1.9"
matplotlib = "^3.9.2"
pillow = "^11.0.0"
wandb = {extras = ["media"], version = "^0.18.5"}
flatdict = "^4.0.1"
seaborn = "^0.13.2"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
